requirements.txt
mysql-connector-python
pandas
numpy
scipy
seaborn